requirements.txt
# Python 3.10 ONLY

# Core LLM stack (legacy)
openai==0.28.1
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.53


# IMPORTANT: legacy langchain needs pydantic v1
pydantic<2

# Config
hydra-core>=1.3.0
omegaconf>=2.3.0

# PDF
pdfplumber>=0.10.2
PyMuPDF>=1.23.2

# Vector DBs
qdrant-client>=1.6.0
chromadb>=0.4.24
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0

# Tools
serpapi @ git+https://github.com/serpapi/serpapi-python.git
requests>=2.31.0

# Utils
python-dotenv>=1.0.0
tiktoken>=0.6.0

# UI
streamlit>=1.35.0

# Dev
ipython>=8.0.0
rich>=13.7.0
networkx

gspread
oauth2client

# Alt LLM
groq>=0.4.0

pandas
matplotlib
seaborn
tabulate
zstandard
